[refactor] 초대 코드 생성 로직을 Code 값객체로 분리 (#48)
